+++ core-ng/src/main/java/core/framework/impl/web/session/RedisSessionStore.java
@@ -1,6 +1,8 @@
 package core.framework.impl.web.session;
 
 import core.framework.api.redis.Redis;
+import core.framework.api.util.JSON;
+import core.framework.api.util.Types;
 
 import java.time.Duration;
 import java.util.Map;
@@ -18,22 +20,28 @@ public class RedisSessionStore implements SessionStore {
     @Override
     public Map<String, String> getAndRefresh(String sessionId, Duration sessionTimeout) {
         String key = sessionKey(sessionId);
-        Map<String, String> data = redis.hash().getAll(key);
-        if (data.isEmpty()) return null;
-
-        redis.expire(key, Duration.ofSeconds(sessionTimeout.getSeconds()));
-        return data;
+        String value = redis.get(key);
+        if (value == null) return null;
+        redis.expire(key, sessionTimeout);
+        return decode(value);
     }
 
     @Override
     public void save(String sessionId, Map<String, String> sessionData, Duration sessionTimeout) {
         String key = sessionKey(sessionId);
-        redis.hash().multiSet(key, sessionData);
-        redis.expire(key, Duration.ofSeconds(sessionTimeout.getSeconds()));
+        redis.set(key, encode(sessionData), sessionTimeout);
+    }
+
+    Map<String, String> decode(String value) {
+        return JSON.fromJSON(Types.map(String.class, String.class), value);
+    }
+
+    String encode(Map<String, String> sessionData) {
+        return JSON.toJSON(sessionData);
     }
 
     @Override
-    public void clear(String sessionId) {
+    public void invalidate(String sessionId) {
         String key = sessionKey(sessionId);
         redis.del(key);
     }