+++ core-ng/src/main/java/core/framework/impl/web/session/SessionImpl.java
@@ -1,6 +1,5 @@
 package core.framework.impl.web.session;
 
-import core.framework.api.util.Exceptions;
 import core.framework.api.util.Maps;
 import core.framework.api.web.Session;
 
@@ -11,7 +10,6 @@ import java.util.Optional;
  * @author neo
  */
 public class SessionImpl implements Session {
-    private static final int MAX_VALUE_LENGTH = 500;
     final Map<String, String> data = Maps.newHashMap();
     String id;
     boolean changed;
@@ -24,9 +22,11 @@ public class SessionImpl implements Session {
 
     @Override
     public void set(String key, String value) {
-        if (value.length() > MAX_VALUE_LENGTH)
-            throw Exceptions.error("the length value must not be larger than {}, length={}", MAX_VALUE_LENGTH, value.length());
-        data.put(key, value);
+        if (value == null) {
+            data.remove(key);
+        } else {
+            data.put(key, value);
+        }
         changed = true;
     }