+++ CHANGELOG.md
@@ -16,6 +16,9 @@
 * mongo: collect mongo stats thru JMX ConnectionPoolStatisticsMBean
 
 ## Change log
+### 4.7.6 (9/13/2016)
+* session: change session data encoding to JSON
+
 ### 4.7.5 (9/7/2016 - 9/8/2016)
 * es: update jackson lib and es lib to match latest
 * json: remove optional field support, which is not useful, only support Optional<T> as return object